+++ examples/standard-use/anthropic-text-image.ts
@@ -7,22 +7,16 @@ config({ path: "../../.env" });
 
 const main = async () => {
   // Create instance
-  const jorEl = new JorEl({
-    vertexAi: {
-      location: process.env.GCP_LOCATION,
-      project: process.env.GCP_PROJECT,
-      keyFilename: process.env.GOOGLE_APPLICATION_CREDENTIALS,
-    },
-  });
+  const jorEl = new JorEl({ anthropic: true });
 
   // Change to vision-capable model
-  jorEl.models.setDefault("gemini-1.5-flash-001");
+  jorEl.models.setDefault("claude-3-5-sonnet-20241022");
 
   // Load image
   const localImage = await ImageContent.fromFile("./image.png");
 
   // Pass image along with the question
-  const response = await jorEl.ask(["Can you describe what is in this image?", localImage]);
+  const response = await jorEl.text(["What is in this image?", localImage]);
 
   console.log(response);
 };