+++ examples/standard-use/groq-text-image.ts
@@ -8,17 +8,18 @@ config({ path: "../../.env" });
 const main = async () => {
   // Create instance
   const jorEl = new JorEl({
-    anthropic: { apiKey: process.env.ANTHROPIC_API_KEY },
+    groq: true,
+    systemMessage: "", // Important for this model to set this to an empty string
   });
 
   // Change to vision-capable model
-  jorEl.models.setDefault("claude-3-5-sonnet-20241022");
+  jorEl.models.setDefault("llama-3.2-11b-vision-preview");
 
   // Load image
   const localImage = await ImageContent.fromFile("./image.png");
 
   // Pass image along with the question
-  const response = await jorEl.ask(["What is in this image?", localImage]);
+  const response = await jorEl.text(["Can you describe what is in this image?", localImage]);
 
   console.log(response);
 };