+++ examples/standard-use/open-ai-text-with-parameters.ts
@@ -6,9 +6,9 @@ import { JorEl } from "../../src";
 config({ path: "../../.env" });
 
 const main = async () => {
-  const jorEl = new JorEl({ openAI: true}); // Uses process.env.OPENAI_API_KEY
+  const jorEl = new JorEl({ openAI: true });
 
-  const response = await jorEl.ask("What is the capital of France?", {
+  const response = await jorEl.text("What is the capital of France?", {
     model: "gpt-4o-mini",
     systemMessage: "You are a helpful assistant",
     temperature: 0.5,