+++ examples/standard-use/open-ai-text.ts
@@ -6,13 +6,13 @@ import { JorEl } from "../../src";
 config({ path: "../../.env" });
 
 const main = async () => {
-  // Create instance with pre-initialized OpenAI provider (and default models)
+  // Create instance
   const jorEl = new JorEl({
-    openAI: { apiKey: process.env.OPENAI_API_KEY },
+    openAI: true,
     systemMessage: "Answer as few words as possible",
   });
 
-  const response = await jorEl.ask("What is the capital of France");
+  const response = await jorEl.text("What is the capital of France");
 
   console.log(response);
   // Paris