+++ examples/standard-use/vertex-ai-text-meta.ts
@@ -7,16 +7,10 @@ config({ path: "../../.env" });
 
 const main = async () => {
   // Create instance
-  const jorEl = new JorEl({
-    vertexAi: {
-      location: process.env.GCP_LOCATION,
-      project: process.env.GCP_PROJECT,
-      keyFilename: process.env.GOOGLE_APPLICATION_CREDENTIALS,
-    },
-  });
+  const jorEl = new JorEl({ vertexAi: true });
 
   // Will return a string
-  const { response, meta } = await jorEl.ask(
+  const { response, meta } = await jorEl.text(
     "What is the capital of France?",
     {
       model: "gemini-1.5-flash-002",